basic server with routing working
